third_party/WebKit/LayoutTests/http/tests/serviceworker/serviceworkerobject-terminate.html

<!DOCTYPE html>
<title>ServiceWorker object: terminate operation</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>

async_test(function(t) {
  var worker = 'resources/empty-worker.js';
  var scope = 'resources/blank.html';
  service_worker_unregister_and_register(t, worker, scope)
    .then(t.step_func(function(registration) {
      return wait_for_update(t, registration);
    }))
    .then(t.step_func(function(worker) {
      assert_throws({name: 'InvalidAccessError'}, function() {
        worker.terminate();
      });
      return service_worker_unregister_and_done(t, scope);
    }))
    .catch(t.step_func(function(e) { throw e; }));
}, 'Verify the terminate operation');

</script>
